# Improve performance when printing large expressions