--- a/components/core/wf/plain_formatter.cc
+++ b/components/core/wf/plain_formatter.cc
@@ -5,7 +5,6 @@
 
 #include "wf/expression_visitor.h"
 #include "wf/utility/assertions.h"
-#include "wf/utility/strings.h"
 #include "wf/utility/third_party_imports.h"
 #include "wf/utility_visitors.h"