Formatting of large expression trees was quite slow, because `plain_formatter` did not use any memoization as it descended through the tree. This change implements caching in the formatter.

The benchmark includes two expression sizes:
- "Small", ~10 operations.
- "Large", ~100 operations composed randomly with deliberate repetitions in the tree.

The performance numbers change as follows:
- Small: `1.65us --> 2.29us`, a 38% regression.
- Large: `2783us --> 150us`, or around a 20x improvement.

In practice, "Large" is going to be more common since most practical expressions _typically_ include many duplicated sub-trees.

As a practical example, the cost of formatting `error_D_rotation_i` in `imu_integration.py` drops from ~143 milliseconds to ~8 milliseconds. This could still be improved meaningfully, although I think a lot of that improvement would be in finding a more elegant way to express really complex expressions.

Parts:
- Add new benchmark on `plain_formatter`.
- Use memoization in `plain_formatter` as the tree is traversed.
- Add better test coverage for formatting of compound expressions to `compound_expression_test.cc`

There is a possibly a **better** alternative implementation of this, where the cache does not contain individual std::string instances but instead contains spans pointing into a single contiguous buffer. This might entail future allocations - for now I am leaving this for future work since the current improvement is already pretty meaningful and there are other areas to focus on.